WebApp/requirements.txt
botbuilder-core
botbuilder-schema
botbuilder-integration-aiohttp
aiohttp